+++ src/index.ts
@@ -3,3 +3,7 @@ export * from "./services/index.js";
 
 // Re-export utilities
 export * from "./utils/index.js";
+
+export type { Country } from "./services/countries/get-countries.js";
+export type { Continent } from "./services/continents/get-continents.js";
+export type { City } from "./services/cities/get-cities.js";